+++ services/streammanager/common/common.go
@@ -4,11 +4,13 @@ package common
 
 import (
 	"encoding/json"
+	"errors"
 	"io"
 	"strings"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/aws/smithy-go"
 )
 
 type StreamProducer interface {
@@ -44,5 +46,38 @@ func ParseAWSError(err error) (statusCode int, respStatus, responseMessage strin
 	}
 
 	statusCode = mapErrorMessageToStatusCode(responseMessage, statusCode)
+	return statusCode, respStatus, responseMessage
+}
+
+func getStatusCodeFromFault(fault smithy.ErrorFault) int {
+	switch fault {
+	case smithy.FaultClient:
+		return 400
+	case smithy.FaultServer:
+		return 500
+	}
+	return 500
+}
+
+func ParseAWSErrorV2(err error) (statusCode int, respStatus, responseMessage string) {
+	statusCode = 500
+	respStatus = "Failure"
+	responseMessage = err.Error()
+
+	var apiErr smithy.APIError
+	if errors.As(err, &apiErr) {
+		responseMessage = apiErr.ErrorMessage()
+		respStatus = apiErr.ErrorCode()
+		fault := apiErr.ErrorFault()
+		statusCode = getStatusCodeFromFault(fault)
+	} else {
+		var opErr *smithy.OperationError
+		if errors.As(err, &opErr) {
+			responseMessage = opErr.Unwrap().Error()
+			statusCode = mapErrorMessageToStatusCode(responseMessage, 400)
+			respStatus = "Failure"
+		}
+	}
+
 	return statusCode, respStatus, responseMessage
 }
