+++ services/streammanager/eventbridge/eventbridgemanager_v2.go
@@ -14,7 +14,7 @@ import (
 	"github.com/rudderlabs/rudder-go-kit/config"
 	"github.com/rudderlabs/rudder-go-kit/jsonrs"
 	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
-	"github.com/rudderlabs/rudder-server/services/streammanager/common"
+	common "github.com/rudderlabs/rudder-server/services/streammanager/common"
 	"github.com/rudderlabs/rudder-server/utils/awsutils"
 )
 
@@ -64,7 +64,7 @@ func (producer *EventBridgeProducerV2) Produce(jsonData json.RawMessage, _ inter
 	// send request to event bridge
 	putEventsOutput, err := client.PutEvents(context.Background(), &requestInput)
 	if err != nil {
-		statusCode, respStatus, responseMessage := common.ParseAWSError(err)
+		statusCode, respStatus, responseMessage := common.ParseAWSErrorV2(err)
 		pkgLogger.Errorf("[EventBridge] error  :: %d : %s : %s", statusCode, respStatus, responseMessage)
 		return statusCode, respStatus, responseMessage
 	}
