+++ services/streammanager/personalize/personalizemanager_v2.go
@@ -85,7 +85,7 @@ func (producer *PersonalizeProducerV2) Produce(jsonData json.RawMessage, _ inter
 	}
 
 	if err != nil {
-		statusCode, respStatus, responseMessage := common.ParseAWSError(err)
+		statusCode, respStatus, responseMessage := common.ParseAWSErrorV2(err)
 		pkgLogger.Errorf("[Personalize] error  :: %d : %s : %s", statusCode, respStatus, responseMessage)
 		return statusCode, respStatus, responseMessage
 	}
